:recycle: Merged reflection utility classes
